+++ spec/testing_record/dsl/builder/settings_spec.rb
@@ -11,6 +11,36 @@ RSpec.describe TestingRecord::DSL::Builder::Settings do
     end
   end
 
+  describe '.attribute' do
+    let(:instance) { refined_class.new }
+
+    context 'when not classified' do
+      let(:refined_class) do
+        Class.new(klazz) do
+          attribute :bar
+        end
+      end
+
+      it 'sets the attribute as a method on the model instance' do
+        expect(instance).to respond_to(:bar)
+      end
+    end
+  end
+
+  describe '.attributes' do
+    context 'when not classified' do
+      let(:refined_class) do
+        Class.new(klazz) do
+          attribute :bar
+        end
+      end
+
+      it 'stores all configured attributes' do
+        expect(refined_class.attributes).to include(:bar)
+      end
+    end
+  end
+
   describe '.caching' do
     before do
       stub_const('SettingsTest', Class.new(klazz))
@@ -26,23 +56,15 @@ RSpec.describe TestingRecord::DSL::Builder::Settings do
     end
   end
 
-  describe '.attribute' do
-    let(:instance) { refined_class.new }
-
-    context 'when not classified' do
-      let(:refined_class) do
-        Class.new(klazz) do
-          attribute :bar
-        end
+  describe '.primary_key' do
+    let(:refined_class) do
+      Class.new(klazz) do
+        primary_key :id
       end
+    end
 
-      it 'sets the attribute as a method on the model instance' do
-        expect(instance).to respond_to(:bar)
-      end
-
-      it 'stores the attribute' do
-        expect(refined_class.attributes).to include(:bar)
-      end
+    it 'stores a primary key value' do
+      expect(refined_class.__primary_key).to eq(:id)
     end
   end
 end